Fix CodeQL alerts: multiplication result converted to larger type
common/argus_client.c (jitter/agr merge statistics, 3 sites): 'float
meanval * int n' was computed at float precision, then widened to a
double destination/accumulator - losing precision that double could
have preserved. Cast meanval to double before the multiply so the
whole computation happens at double precision.

examples/ramysql/rascore.c, examples/ramysql/rasql.c,
examples/ratop/raclient.c (RaSQLProcessQueue/equivalent, time bucketing):
same 'secs = (fstruct->second/RaPeriod) * RaPeriod' pattern already
fixed in examples/rabehave/raclient.c and examples/ramysql/rasrvs.c
during earlier triage - cast fstruct->second (unsigned int) to time_t
before the divide so the result type matches the destination.

examples/ramatrix/ramatrix.c (RaMatrixCalculate-equivalent matrix
allocation): 'n * n' computed in int before being passed as a size_t
calloc() argument - can overflow int when queue->count is large
enough, before ever reaching the size_t parameter. Cast to size_t
before the multiply.

Verified: full clean rebuild (0 errors, 0 warnings), smoke test against
dump_dir_2_thinwin_rdp.argus.out unchanged (ra/radump/radns: 44/44/10 lines).

+++ common/argus_client.c
@@ -7542,7 +7542,7 @@ ArgusMergeRecords (const struct ArgusAggregatorStruct * const na,
                                  if (a1->act.minval > value)
                                     a1->act.minval = value;
 
-                              sum1  = a1->act.meanval * a1->act.n;
+                              sum1  = (double)a1->act.meanval * a1->act.n;
                               sum1 += value;
 
                               if (a1->act.stdev != 0) {
@@ -7708,7 +7708,7 @@ ArgusMergeRecords (const struct ArgusAggregatorStruct * const na,
                                           ((double) j2->dst.act.meanval * (double) j2->dst.act.n)) / n;
 
                               if (j1->dst.act.n) {
-                                 double sum  =  j1->dst.act.meanval * j1->dst.act.n;
+                                 double sum  =  (double)j1->dst.act.meanval * j1->dst.act.n;
                                  sumsqrd += (j1->dst.act.n * ((double)j1->dst.act.stdev * (double)j1->dst.act.stdev)) +
                                             (sum * sum)/j1->dst.act.n;
                               }
@@ -8594,7 +8594,7 @@ ArgusIntersectRecords (struct ArgusAggregatorStruct *na, struct ArgusRecordStruc
                                  ((double) j2->dst.act.meanval * (double) j2->dst.act.n)) / n;
 
                      if (j1->dst.act.n) {
-                        double sum  =  j1->dst.act.meanval * j1->dst.act.n;
+                        double sum  =  (double)j1->dst.act.meanval * j1->dst.act.n;
                         sumsqrd += (j1->dst.act.n * ((double)j1->dst.act.stdev * (double)j1->dst.act.stdev)) +
                                    (sum * sum)/j1->dst.act.n;
                      }
